[refactor] Reorganize collections crate
