+++ crates/samlang-collections/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.8.0"
 edition = "2021"
 
 [dependencies]
-itertools = "0.11.0"
+fplist = "0.2.1"
 
 [dev-dependencies]
 pretty_assertions = "1.4.0"
