+++ crates/samlang-core/src/optimization/local_value_numbering.rs
@@ -1,6 +1,6 @@
 use super::optimization_common::{BinaryBindedValue, BindedValue, IndexAccessBindedValue};
 use crate::ast::mir::{Binary, Callee, Expression, Function, Statement, VariableName};
-use samlang_collections::LocalStackedContext;
+use samlang_collections::local_stacked_context::LocalStackedContext;
 use samlang_heap::PStr;
 
 type LocalContext = LocalStackedContext<PStr, PStr>;
